# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Access Audit API
  version: 1.0.0
  description: >
    Recruiter-ready API for identity governance, privileged access analysis,
    audit visibility, and remediation-aware access review workflows.
servers:
  - url: http://localhost:3000
paths:
...
  /api/analyze/access-risk:
    post:
      summary: Analyze access risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Risk analysis
  /api/analyze/exception:
    post:
      summary: Analyze policy exception posture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Exception analysis
  /api/analyze/remediation:
    post:
      summary: Analyze remediation priority
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Remediation analysis